pyproject.toml
[tool.poetry]
name = "confighole"
version = "1.0.0"
description = "Pi-hole configuration as code"
authors = ["dsgnr"]
readme = "README.md"
homepage = "https://github.com/dsgnr/confighole"
repository = "https://github.com/dsgnr/confighole"
documentation = "https://github.com/dsgnr/confighole"
keywords = ["pihole", "dns", "configuration", "management"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Systems Administration",
]
packages = [{include = "confighole"}]

[tool.poetry.dependencies]
python = "^3.13"
pihole-lib = {git = "https://github.com/dsgnr/pihole-lib.git"}
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
mypy = "^1.7.0"
types-pyyaml = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
docker = "^7.0.0"
requests = "^2.31.0"

[tool.poetry.scripts]
confighole = "confighole.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88  # Black's default line length
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
# Black-compatible settings
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.isort]
known-first-party = ["confighole"]
# Black-compatible isort settings
combine-as-imports = true
split-on-trailing-comma = true
force-single-line = false

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "pihole_lib.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["confighole"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]